\name{Rserve}
\title{Server providing R functionality to applications via TCP/IP}
\alias{Rserve}
\usage{
R CMD Rserve
}
\description{
Starts R server in daemon mode (unix only). Any additional parameters will be passed staright to the underlying R.
}
\arguments{
}
\details{
Rserve is not a library, but an application. It is provided an A R package for convenience only. For details see http://www.rosuda.org/Rserve
}
\author{S. Urbanek}
\keyword{Rserve}
